+++ test/smoke/testbunchofsousservers_test.go
@@ -104,6 +104,7 @@ func (c *TestBunchOfSousServers) Configure(t *testing.T, envDesc desc.EnvDesc) e
 			StateLocation: i.StateDir,
 			SiblingURLs:   siblingURLs,
 			Database: storage.PostgresConfig{
+				User:   "postgres",
 				DBName: dbname,
 				Host:   "localhost",
 				Port:   dbport,