+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 with respect to its command line interface and HTTP interface
 
 ## [Unreleased](//github.com/opentable/sous/compare/0.5.91...HEAD)
+### Changed
 * Server: Postgres is now required for the server to operate.
 * Both: SQLite is removed as a dependency of Sous.
+
+### Added
+* Client: Deploy with zero instances will give a specific error message that you have zero instances
 
 ## [0.5.91](//github.com/opentable/sous/compare/0.5.88...0.5.91)
 ### Added